requirements.txt
flask
requests
ics
pycryptodome
python-dateutil
pytz
rfc3339
